+++ src/webportal/src/plugins/job-submission/components/App.jsx
@@ -1,3 +1,28 @@
+/*!
+ * Copyright (c) Microsoft Corporation
+ * All rights reserved.
+ *
+ * MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 import React from 'react';
 import { Customizer, Stack } from 'office-ui-fabric-react';
 import { FluentCustomizations } from '@uifabric/fluent-theme';
@@ -7,7 +32,6 @@ import { TabFormContent } from './TabFormContent';
 import { JobTaskRole } from '../models/jobTaskRole';
 import { JobInformation } from './JobInformation';
 import { Parameter } from './Parameter';
-
 
 initializeIcons();
 
@@ -23,7 +47,8 @@ export class App extends React.Component {
 
   _onItemAdd() {
     const { items } = this.state;
-    const updatedItems = [...items, {headerText: 'Task role X', content: new JobTaskRole()}];
+    // TODO: This just for test, need to correct the header text
+    const updatedItems = [...items, {headerText: `Task role ${items.length + 1}`, content: new JobTaskRole()}];
     this.setState({
       items: updatedItems
     });
