+++ src/webportal/src/plugins/job-submission/components/BasicSection.jsx
@@ -1,3 +1,28 @@
+/*!
+ * Copyright (c) Microsoft Corporation
+ * All rights reserved.
+ *
+ * MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 import React, { useState } from 'react';
 import { getId } from 'office-ui-fabric-react/lib/Utilities';
 import { Label, Stack, Text, Icon } from 'office-ui-fabric-react';
@@ -13,15 +38,15 @@ export const BasicSection = (props) => {
   const {label, optional, children } = props;
   const textFieldId = getId('textField');
 
-  const [isItemOn, setIsItemOn] = useState(false);
+  const [isSectionOn, setSectionOn] = useState(false);
   const [iconName, setIconName] = useState('CaretSolidRight');
   const onOpenItem = () => {
-    if (isItemOn) {
+    if (isSectionOn) {
       setIconName('CaretSolidRight');
     } else {
       setIconName('CaretDown8');
     }
-    setIsItemOn(!isItemOn);
+    setSectionOn(!isSectionOn);
   };
 
   return (
@@ -36,7 +61,7 @@ export const BasicSection = (props) => {
       </FormColumn>
       <FormColumn styles={formPageStyle.formSecondColunm}>
         <Stack gap={marginSize.s1}>
-          {(!optional || isItemOn) && children}
+          {(!optional || isSectionOn) && children}
         </Stack>
       </FormColumn>
     </FormSection>
