+++ src/webportal/src/plugins/job-submission/index.js
@@ -22,6 +22,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+
+ // TODO: This file need to change in the future. Submit job may no be a plugin
 import React from "react";
 import ReactDOM from "react-dom";
